More object orientation, please
