+++ templates/person/view.html
@@ -84,56 +84,7 @@
     {% if memberships %}
     <div class="row">
         <div class="col-sm-12">
-            <h3 id="memberships">
-                <i class="fa fa-fw fa-users"></i>
-                {% trans "Memberships" %}
-                <small class="pull-right">{% help href='whowasincommand/personsrec.html#person_record_anatomy_memberships' %}</small>
-            </h3>
-            <hr />
-            <table class="table table-condensed">
-                <thead>
-                    <tr>
-                        <th>{% trans "Organization" %}</th>
-                        <th>{% trans "Rank" %}</th>
-                        <th>{% trans "Role" %}</th>
-                        <th>{% trans "Title" %}</th>
-                        <th>{% trans "First cited" %}</th>
-                        <th>{% trans "Last cited" %}</th>
-                    </tr>
-                </thead>
-                <tbody>
-                    {% for member in memberships %}
-                        <tr>
-                            <td class="cited">
-                                <a href="{% url 'view-organization' member.organization.get_value.value.id %}">
-                                    {{ member.organization.get_value }}
-                                </a>
-                                {% cite member.organization.get_value %}
-                            </td>
-                            <td class="cited">
-                                {{ member.rank.get_value|default:"" }}
-                                {% cite member.rank.get_value %}
-                            </td>
-                            <td class="cited">
-                                {{ member.role.get_value|default:"" }}
-                                {% cite member.role.get_value %}
-                            </td>
-                            <td class="cited">
-                                {{ member.title.get_value|default:"" }}
-                                {% cite member.title.get_value %}
-                            </td>
-                            <td class="cited">
-                                {% datetype member.firstciteddate.get_value 'start' %}
-                                {% cite member.firstciteddate.get_value %}
-                            </td>
-                            <td class="cited">
-                                {% datetype member.lastciteddate.get_value 'end' %}
-                                {% cite member.lastciteddate.get_value %}
-                            </td>
-                        </tr>
-                    {% endfor %}
-                </tbody>
-            </table>
+            {% include 'person/partials/person-posting-list.html' %}
         </div>
     </div>
     {% endif %}
